refactor(fgumi-sort): move sort/ from main crate (move + transplant)
Moves all 16 files from src/lib/sort/ to crates/fgumi-sort/src/ and absorbs
the original src/lib/sort/mod.rs into crates/fgumi-sort/src/lib.rs.

Earlier commits on this branch were strictly "move-only" git mv renames.
This one cannot be: the new crate's lib.rs already existed as a 10-line
Phase A scaffold, so git cannot detect a rename when sort/mod.rs's body
replaces it. The numstat shows +226/-8 on lib.rs (and -228 on the deleted
mod.rs), but the post-move lib.rs body is byte-identical to the pre-move
sort/mod.rs body — verified by diffing `<this-commit>:crates/fgumi-sort/src/lib.rs`
against `<this-commit>^:src/lib/sort/mod.rs`.

The "(move-only)" tag on the original subject was misleading; "(move +
transplant)" describes the operation accurately.
