+++ crates/fgumi-sort/src/inline_buffer.rs
