+++ crates/fgumi-sort/src/pipeline.rs
