+++ crates/fgumi-sort/src/pooled_bam_writer.rs
