+++ crates/fgumi-sort/src/segmented_buf.rs
